Split huge DoxygenWindow file into 4
