+++ Editor/DoxygenWindow/DoxygenConfig.cs
@@ -1,0 +1,36 @@
+/*
+Original copyright notice by Jacob Pennock:
+
+/// <summary>
+/// <para> A Editor Plugin for automatic doc generation through Doxygen</para>
+/// <para> Author: Jacob Pennock (http://Jacobpennock.com)</para>
+/// <para> Version: 1.0</para>	 
+/// </summary>
+
+Permission is hereby granted, free of charge, to any person  obtaining a copy of this software and associated documentation  files (the "Software"), to deal in the Software without  restriction, including without limitation the rights to use,  copy, modify, merge, publish, distribute, sublicense, and/or sell  copies of the Software, and to permit persons to whom the  Software is furnished to do so, subject to the following conditions:
+The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*/
+
+
+using System;
+using System.Collections.Generic;
+
+namespace GuiToolkit.Editor
+{
+	/// <summary> 
+	/// <para>A small data structure class hold values for making Doxygen config files </para>
+	/// </summary>
+	[Serializable]
+	public class DoxygenConfig
+	{
+		public string Project;
+		public string Synopsis = "";
+		public string Version = "";
+		public string ScriptsDirectory;
+		public string DocDirectory;
+		public string PathtoDoxygen = "";
+		public string Defines;
+		public List<string> ExcludePatterns;
+	}
+}
