+++ Editor/DoxygenWindow/DoxygenThreadSafeOutput.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3c52ae0d59de3ff4db58506f300b8bba
