+++ Editor/DoxygenWindow/DoxygenWindow.cs
@@ -1,6 +1,12 @@
 /*
 Original copyright notice by Jacob Pennock:
 
+/// <summary>
+/// <para> A Editor Plugin for automatic doc generation through Doxygen</para>
+/// <para> Author: Jacob Pennock (http://Jacobpennock.com)</para>
+/// <para> Version: 1.0</para>	 
+/// </summary>
+
 Permission is hereby granted, free of charge, to any person  obtaining a copy of this software and associated documentation  files (the "Software"), to deal in the Software without  restriction, including without limitation the rights to use,  copy, modify, merge, publish, distribute, sublicense, and/or sell  copies of the Software, and to permit persons to whom the  Software is furnished to do so, subject to the following conditions:
 The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
@@ -10,36 +16,16 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 using UnityEngine;
 using UnityEditor;
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Threading;
 using System.Text;
-using System.Text.RegularExpressions;
 
 namespace GuiToolkit.Editor
 {
-	/// <summary> 
-	/// <para>A small data structure class hold values for making Doxygen config files </para>
-	/// </summary>
-	[Serializable]
-	public class DoxygenConfig
-	{
-		public string Project;
-		public string Synopsis = "";
-		public string Version = "";
-		public string ScriptsDirectory;
-		public string DocDirectory;
-		public string PathtoDoxygen = "";
-		public string Defines;
-		public List<string> ExcludePatterns;
-	}
-
 	/// <summary>
-	/// <para> A Editor Plugin for automatic doc generation through Doxygen</para>
-	/// <para> Author: Jacob Pennock (http://Jacobpennock.com)</para>
-	/// <para> Version: 1.0</para>	 
+	/// <para> An Editor Plugin for automatic doc generation through Doxygen</para>
+	/// <para> Original Author: Jacob Pennock (http://Jacobpennock.com)</para>
 	/// </summary>
 	public class DoxygenWindow : EditorWindow
 	{
@@ -67,7 +53,7 @@ namespace GuiToolkit.Editor
 		public string BaseFileString = null;
 		public string DoxygenOutputString = null;
 		public string CurentOutput = null;
-		DoxyThreadSafeOutput DoxygenOutput = null;
+		DoxygenThreadSafeOutput DoxygenOutput = null;
 		List<string> DoxygenLog = null;
 		bool ViewLog = false;
 		Vector2 scroll;
@@ -77,7 +63,7 @@ namespace GuiToolkit.Editor
 		[MenuItem("Window/Documentation with Doxygen")]
 		public static void Init()
 		{
-			Instance = (DoxygenWindow)EditorWindow.GetWindow(typeof(DoxygenWindow), false, "Documentation");
+			Instance = (DoxygenWindow)GetWindow(typeof(DoxygenWindow), false, "Documentation");
 			Instance.minSize = new Vector2(420, 245);
 			Instance.maxSize = new Vector2(420, 720);
 
@@ -85,7 +71,7 @@ namespace GuiToolkit.Editor
 
 		void OnEnable()
 		{
-			UnityProjectID = PlayerSettings.productName + ":";
+			UnityProjectID = UnityEditor.PlayerSettings.productName + ":";
 			AssetsFolder = Application.dataPath;
 			LoadConfig();
 			DoxyoutputProgress = 0;
@@ -488,7 +474,7 @@ namespace GuiToolkit.Editor
 			if (!LoadSavedConfig())
 			{
 				Config = new DoxygenConfig();
-				Config.Project = PlayerSettings.productName;
+				Config.Project = UnityEditor.PlayerSettings.productName;
 				Config.ScriptsDirectory = Application.dataPath;
 				Config.DocDirectory = Application.dataPath.Replace("Assets", "Docs");
 			}
@@ -596,312 +582,16 @@ namespace GuiToolkit.Editor
 			string[] Args = new string[1];
 			Args[0] = Config.DocDirectory + "/Doxyfile";
 
-			DoxygenOutput = new DoxyThreadSafeOutput();
+			DoxygenOutput = new DoxygenThreadSafeOutput();
 			DoxygenOutput.SetStarted();
 
 			Action<int> setcallback = (int returnCode) => OnDoxygenFinished(returnCode);
 
-			DoxyRunner Doxygen = new DoxyRunner(Config.PathtoDoxygen, Args, DoxygenOutput, setcallback);
+			DoxygenRunner Doxygen = new DoxygenRunner(Config.PathtoDoxygen, Args, DoxygenOutput, setcallback);
 
 			Thread DoxygenThread = new Thread(new ThreadStart(Doxygen.RunThreadedDoxy));
 			DoxygenThread.Start();
 		}
 
 	}
-
-	/// <summary>
-	///  This class spawns and runs Doxygen in a separate thread, and could serve as an example of how to create 
-	///  plugins for unity that call a command line application and then get the data back into Unity safely.	 
-	/// </summary>
-	public class DoxyRunner
-	{
-		DoxyThreadSafeOutput SafeOutput;
-		public Action<int> onCompleteCallBack;
-		List<string> DoxyLog = new List<string>();
-		public string EXE = null;
-		public string[] Args;
-		static string WorkingFolder;
-
-		public DoxyRunner(string exepath, string[] args, DoxyThreadSafeOutput safeoutput, Action<int> callback)
-		{
-			EXE = exepath;
-			Args = args;
-			SafeOutput = safeoutput;
-			onCompleteCallBack = callback;
-			WorkingFolder = exepath.Replace("doxygen.exe", "");
-		}
-
-		public void updateOuputString(string output)
-		{
-			SafeOutput.WriteLine(output);
-			DoxyLog.Add(output);
-		}
-
-		public void RunThreadedDoxy()
-		{
-			Action<string> GetOutput = (string output) => updateOuputString(output);
-			int ReturnCode = Run(GetOutput, null, EXE, Args);
-			SafeOutput.WriteFullLog(DoxyLog);
-			SafeOutput.SetFinished();
-			onCompleteCallBack(ReturnCode);
-		}
-
-		/// <summary>
-		/// Runs the specified executable with the provided arguments and returns the process' exit code.
-		/// </summary>
-		/// <param name="output">Recieves the output of either std/err or std/out</param>
-		/// <param name="input">Provides the line-by-line input that will be written to std/in, null for empty</param>
-		/// <param name="exe">The executable to run, may be unqualified or contain environment variables</param>
-		/// <param name="args">The list of unescaped arguments to provide to the executable</param>
-		/// <returns>Returns process' exit code after the program exits</returns>
-		/// <exception cref="System.IO.FileNotFoundException">Raised when the exe was not found</exception>
-		/// <exception cref="System.ArgumentNullException">Raised when one of the arguments is null</exception>
-		/// <exception cref="System.ArgumentOutOfRangeException">Raised if an argument contains '\0', '\r', or '\n'
-		public static int Run(Action<string> output, TextReader input, string exe, params string[] args)
-		{
-			if (String.IsNullOrEmpty(exe))
-				throw new FileNotFoundException();
-			if (output == null)
-				throw new ArgumentNullException("output");
-
-			ProcessStartInfo psi = new ProcessStartInfo();
-			psi.UseShellExecute = false;
-			psi.RedirectStandardError = true;
-			psi.RedirectStandardOutput = true;
-			psi.RedirectStandardInput = true;
-			psi.WindowStyle = ProcessWindowStyle.Hidden;
-			psi.CreateNoWindow = true;
-			psi.ErrorDialog = false;
-			psi.WorkingDirectory = WorkingFolder;
-			psi.FileName = FindExePath(exe);
-			psi.Arguments = EscapeArguments(args);
-
-			using (Process process = Process.Start(psi))
-			using (ManualResetEvent mreOut = new ManualResetEvent(false),
-			       mreErr = new ManualResetEvent(false))
-			{
-				process.OutputDataReceived += (o, e) =>
-				{
-					if (e.Data == null) mreOut.Set();
-					else output(e.Data);
-				};
-				process.BeginOutputReadLine();
-				process.ErrorDataReceived += (o, e) =>
-				{
-					if (e.Data == null) mreErr.Set();
-					else output(e.Data);
-				};
-				process.BeginErrorReadLine();
-
-				string line;
-				while (input != null && null != (line = input.ReadLine()))
-					process.StandardInput.WriteLine(line);
-
-				process.StandardInput.Close();
-				process.WaitForExit();
-
-				mreOut.WaitOne();
-				mreErr.WaitOne();
-				return process.ExitCode;
-			}
-		}
-
-		/// <summary>
-		/// Quotes all arguments that contain whitespace, or begin with a quote and returns a single
-		/// argument string for use with Process.Start().
-		/// </summary>
-		/// <param name="args">A list of strings for arguments, may not contain null, '\0', '\r', or '\n'</param>
-		/// <returns>The combined list of escaped/quoted strings</returns>
-		/// <exception cref="System.ArgumentNullException">Raised when one of the arguments is null</exception>
-		/// <exception cref="System.ArgumentOutOfRangeException">Raised if an argument contains '\0', '\r', or '\n'</exception>
-		public static string EscapeArguments(params string[] args)
-		{
-			StringBuilder arguments = new StringBuilder();
-			Regex invalidChar = new Regex("[\x00\x0a\x0d]"); //  these can not be escaped
-			Regex needsQuotes = new Regex(@"\s|"""); //          contains whitespace or two quote characters
-			Regex escapeQuote = new Regex(@"(\\*)(""|$)"); //    one or more '\' followed with a quote or end of string
-			for (int carg = 0; args != null && carg < args.Length; carg++)
-			{
-				if (args[carg] == null)
-				{
-					throw new ArgumentNullException("args[" + carg + "]");
-				}
-
-				if (invalidChar.IsMatch(args[carg]))
-				{
-					throw new ArgumentOutOfRangeException("args[" + carg + "]");
-				}
-
-				if (args[carg] == String.Empty)
-				{
-					arguments.Append("\"\"");
-				}
-				else if (!needsQuotes.IsMatch(args[carg]))
-				{
-					arguments.Append(args[carg]);
-				}
-				else
-				{
-					arguments.Append('"');
-					arguments.Append(escapeQuote.Replace(args[carg], m =>
-						m.Groups[1].Value + m.Groups[1].Value +
-						(m.Groups[2].Value == "\"" ? "\\\"" : "")
-					));
-					arguments.Append('"');
-				}
-
-				if (carg + 1 < args.Length)
-					arguments.Append(' ');
-			}
-
-			return arguments.ToString();
-		}
-
-
-		/// <summary>
-		/// Expands environment variables and, if unqualified, locates the exe in the working directory
-		/// or the evironment's path.
-		/// </summary>
-		/// <param name="exe">The name of the executable file</param>
-		/// <returns>The fully-qualified path to the file</returns>
-		/// <exception cref="System.IO.FileNotFoundException">Raised when the exe was not found</exception>
-		public static string FindExePath(string exe)
-		{
-			exe = Environment.ExpandEnvironmentVariables(exe);
-			if (!File.Exists(exe))
-			{
-				if (Path.GetDirectoryName(exe) == String.Empty)
-				{
-					foreach (string test in (Environment.GetEnvironmentVariable("PATH") ?? "").Split(';'))
-					{
-						string path = test.Trim();
-						if (!String.IsNullOrEmpty(path) && File.Exists(path = Path.Combine(path, exe)))
-							return Path.GetFullPath(path);
-					}
-				}
-
-				throw new FileNotFoundException(new FileNotFoundException().Message, exe);
-			}
-
-			return Path.GetFullPath(exe);
-		}
-	}
-
-
-	/// <summary>
-	///  This class encapsulates the data output by Doxygen so it can be shared with Unity in a thread share way.	 
-	/// </summary>
-	public class DoxyThreadSafeOutput
-	{
-		private ReaderWriterLockSlim outputLock = new ReaderWriterLockSlim();
-		private string CurrentOutput = "";
-		private List<string> FullLog = new List<string>();
-		private bool Finished = false;
-		private bool Started = false;
-
-		public string ReadLine()
-		{
-			outputLock.EnterReadLock();
-			try
-			{
-				return CurrentOutput;
-			}
-			finally
-			{
-				outputLock.ExitReadLock();
-			}
-		}
-
-		public void SetFinished()
-		{
-			outputLock.EnterWriteLock();
-			try
-			{
-				Finished = true;
-			}
-			finally
-			{
-				outputLock.ExitWriteLock();
-			}
-		}
-
-		public void SetStarted()
-		{
-			outputLock.EnterWriteLock();
-			try
-			{
-				Started = true;
-			}
-			finally
-			{
-				outputLock.ExitWriteLock();
-			}
-		}
-
-		public bool isStarted()
-		{
-			outputLock.EnterReadLock();
-			try
-			{
-				return Started;
-			}
-			finally
-			{
-				outputLock.ExitReadLock();
-			}
-		}
-
-		public bool isFinished()
-		{
-			outputLock.EnterReadLock();
-			try
-			{
-				return Finished;
-			}
-			finally
-			{
-				outputLock.ExitReadLock();
-			}
-		}
-
-		public List<string> ReadFullLog()
-		{
-			outputLock.EnterReadLock();
-			try
-			{
-				return FullLog;
-			}
-			finally
-			{
-				outputLock.ExitReadLock();
-			}
-		}
-
-		public void WriteFullLog(List<string> newLog)
-		{
-			outputLock.EnterWriteLock();
-			try
-			{
-				FullLog = newLog;
-			}
-			finally
-			{
-				outputLock.ExitWriteLock();
-			}
-		}
-
-		public void WriteLine(string newOutput)
-		{
-			outputLock.EnterWriteLock();
-			try
-			{
-				CurrentOutput = newOutput;
-			}
-			finally
-			{
-				outputLock.ExitWriteLock();
-			}
-		}
-	}
-
 }
